pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sessionbus"
version = "0.1.0"
description = "Local-first Session I/O Bus for coordinating AI agent sessions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.111.0,<1.0.0",
  "uvicorn>=0.30.0,<1.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "aiosqlite>=0.20.0,<1.0.0",
  "jinja2>=3.1.4,<4.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic>=2.7.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]
desktop = [
  "pywebview>=5.1,<6.0",
]
mcp = [
  "mcp>=1.0.0,<2.0.0",
]
all = [
  "pytest>=8.2.0",
  "pywebview>=5.1,<6.0",
  "mcp>=1.0.0,<2.0.0",
]

[project.scripts]
agent-flow = "sessionbus.main:cli"
sessionbus-hub = "sessionbus.hub:cli"
sessionbus-mcp = "sessionbus.mcp_service:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["sessionbus*"]

[tool.setuptools.package-data]
sessionbus = ["templates/*.html", "static/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
